refactor: transform cjs into mjs module
