+++ src/modules/files/b.json
@@ -1,5 +1,5 @@
 {
-    "a": 11,
-    "b": 22,
-    "c": 33
+  "a": 11,
+  "b": 22,
+  "c": 33
 }
